+++ js/checker.js
@@ -1,0 +1,1 @@
+/*Сравнение похожих объявлений*/
