+++ js/validator.js
@@ -1,0 +1,1 @@
+/*Будущая валидация форм*/
